WS_TK_Ticket_Download.java WS_TK_Ticket_Search.java     - Tratativa para impedir que ticket com form espontaneo em processo seja atualizado pelo server. @#$ Act011_Main.java     - Correção na navegação ao delete form. @#$ Act011_Main_Presenter_Impl.java     - Metodo para identificação de form espontaneo do ticket.     - Adição de Step_code na query de delete de Ctrl.     - Correção de metodo que identifica se é um formulario oriundo do ticket(planned ou nao).     - Correção da informações de produto e serial no ctrl. @#$ Act070_Main.java     - Trava de sincronismo quando houver form espontaneo em processo. @#$ Act070_Main_Presenter.java     - Trava de sincronismo quando houver form espontaneo em processo.     - Correção no acesso de forms espontaneos. @#$ Act071_Main_Presenter.java     - Finaliza action offline caso haja form espontaneo em processo. @#$ ToolBox_Inf.java     - Metodo que verifica se há algum form espontaneo em processo para o ticket.     - Troca de query que contabiliza tickets pendentes de envio. @#$ Sql_Act068_001.java Sql_WS_TK_Ticket_Save_001.java     - Adição de condição de formulario espontaneo em processo. @#$ TK_Ticket_Ctrl_Sql_005.java     - Adição de step_code para garantir delete unico. @#$ TK_Ticket_Ctrl_Sql_006.java     - COrreção para seleção correta de ctrl de form. @#$

+++ app/src/main/java/com/namoadigital/prj001/service/WS_TK_Ticket_Download.java
@@ -150,35 +150,47 @@ public class WS_TK_Ticket_Download extends IntentService {
             DaoObjReturn daoObjReturn = new DaoObjReturn();
             //
             HMAux hmAux = new HMAux();
+            List<TK_Ticket> tickets = new ArrayList<>();
             for (TK_Ticket tkTicket : ticketList) {
                 tkTicket.setPK();
                 TK_Ticket dbTicket = getDbTicket(tkTicket);
+
                 if(dbTicket != null) {
-                    //Verifica se precisa resetar alguma foto. Isso deve ser feito se o "file_code" da foto
-                    //for alterado, o que significa que mudaram a foto no server...
-                    TK_Ticket.checkActionPhotoResetNeeds(
-                        dbTicket,
-                        tkTicket
-                    );
-                    //Varre todas as imagens verificando se existe imagem local para cada item que pode ter foto
-                    tkTicket.updateLocalImagesPathIfExists();
-                    //Busca ctrls tipo form em andamento e que seriam resetados.
-                    tkTicket.updateTicketCtrlFormInProcess(getApplicationContext());
-                    //
-                    daoObjReturn = ticketDao.removeFullV2(tkTicket);
+                    /*
+                        Barrionuevo - 2020-11-13
+                        Tratativa para impedir que ticket com form espontaneo em processo seja atualizado pelo server.
+                     */
+                    if(!ToolBox_Inf.hasOffHandFormInProcess(getApplicationContext(), dbTicket.getTicket_prefix(), dbTicket.getTicket_code())) {
+                        tickets.add(dbTicket);
+
+                        //Verifica se precisa resetar alguma foto. Isso deve ser feito se o "file_code" da foto
+                        //for alterado, o que significa que mudaram a foto no server...
+                        TK_Ticket.checkActionPhotoResetNeeds(
+                                dbTicket,
+                                tkTicket
+                        );
+                        //Varre todas as imagens verificando se existe imagem local para cada item que pode ter foto
+                        tkTicket.updateLocalImagesPathIfExists();
+                        //Busca ctrls tipo form em andamento e que seriam resetados.
+                        tkTicket.updateTicketCtrlFormInProcess(getApplicationContext());
+                        //
+                        daoObjReturn = ticketDao.removeFullV2(tkTicket);
+                    }
                 }
                 //
                 if(daoObjReturn.hasError()) {
                     break;
-                }else {
-                    if (ticketList.size() == 1) {
-                        hmAux.put(TK_TicketDao.TICKET_PREFIX, String.valueOf(tkTicket.getTicket_prefix()));
-                        hmAux.put(TK_TicketDao.TICKET_CODE, String.valueOf(tkTicket.getTicket_code()));
-                    }
                 }
             }
+            //
+            if (tickets.size() == 1) {
+                TK_Ticket tkTicket = tickets.get(0);
+                hmAux.put(TK_TicketDao.TICKET_PREFIX, String.valueOf(tkTicket.getTicket_prefix()));
+                hmAux.put(TK_TicketDao.TICKET_CODE, String.valueOf(tkTicket.getTicket_code()));
+            }
+            //
             if(!daoObjReturn.hasError()) {
-                daoObjReturn = ticketDao.addUpdate(ticketList, false);
+                daoObjReturn = ticketDao.addUpdate(tickets, false);
                 if(!daoObjReturn.hasError()){
                     startDownloadServices();
                     //
